Fix drop_last heisenbug
When the sequence length is not uniform,
categories do not all have the same number
of samples. This can make certain categories get
dropped if they are in an unfortunate position in
the indices list.

The net effect is to blow up some tests, randomly.
Since this behaviour is known (incomplete batches cannot
be used), it is not reasonable to fail tests when
it is encountered. Thus all sequences are given the
same length for the time being.